Convert news to App Router
