+++ demo/site/src/news/NewsList.fragment.ts
@@ -1,0 +1,13 @@
+import { gql } from "graphql-request";
+
+export const newsListFragment = gql`
+    fragment NewsList on PaginatedNews {
+        nodes {
+            id
+            title
+            slug
+            image
+            createdAt
+        }
+    }
+`;
